move config value factory into repo layer
