+++ apps/server/src/infra/runtime-config/repo/runtime-config-value.factory.ts
@@ -1,4 +1,4 @@
-import { RuntimeConfigValue, RuntimeConfigValueProps } from './runtime-config-value.do';
+import { RuntimeConfigValue, RuntimeConfigValueProps } from '../domain/runtime-config-value.do';
 
 export class RuntimeConfigValueFactory {
 	public static build(props: RuntimeConfigValueProps): RuntimeConfigValue {
